requirements.txt
anyio==3.6.1; python_version >= "3.6" and python_full_version >= "3.6.2"
asyncpg==0.26.0; python_full_version >= "3.6.0"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
dnspython==2.2.1; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.1"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
email-validator==1.2.1; python_full_version >= "3.6.1"
fastapi==0.81.0; python_full_version >= "3.6.1"
greenlet==1.1.3; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.13.0; python_version >= "3.7"
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.6"
passlib==1.7.4
psycopg2-binary==2.9.3; python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pydantic==1.9.2; python_full_version >= "3.6.1"
python-dotenv==0.21.0; python_version >= "3.7"
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rsa==4.9; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
sqlalchemy==1.4.40; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.3.0; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "3.10"
uvicorn==0.18.3; python_version >= "3.7"
